refactoring torrent search/selection part
